+++ .github/ISSUE_TEMPLATE/config.yml
@@ -3,4 +3,4 @@ blank_issues_enabled: false
 contact_links:
   - name: MMSegmentation Documentation
     url: https://mmsegmentation.readthedocs.io
-    about: Check the docs and FAQ to see if you question is already anwsered.
+    about: Check the docs and FAQ to see if you question is already answered.
