+++ .github/ISSUE_TEMPLATE/error-report.md
@@ -30,7 +30,7 @@ A clear and concise description of what the bug is.
 
 **Environment**
 
-1. Please run `python mmseg/utils/collect_env.py` to collect necessary environment infomation and paste it here.
+1. Please run `python mmseg/utils/collect_env.py` to collect necessary environment information and paste it here.
 2. You may add addition that may be helpful for locating the problem, such as
     - How you installed PyTorch [e.g., pip, conda, source]
     - Other environment variables that may be related (such as `$PATH`, `$LD_LIBRARY_PATH`, `$PYTHONPATH`, etc.)
