+++ docs/tutorials/customize_runtime.md
@@ -113,7 +113,7 @@ Tricks not implemented by the optimizer should be implemented through optimizer 
         _delete_=True, grad_clip=dict(max_norm=35, norm_type=2))
     ```
 
-    If your config inherits the base config which already sets the `optimizer_config`, you might need `_delete_=True` to overide the unnecessary settings. See the [config documenetation](https://mmsegmentation.readthedocs.io/en/latest/config.html) for more details.
+    If your config inherits the base config which already sets the `optimizer_config`, you might need `_delete_=True` to override the unnecessary settings. See the [config documentation](https://mmsegmentation.readthedocs.io/en/latest/config.html) for more details.
 
 - __Use momentum schedule to accelerate model convergence__:
     We support momentum scheduler to modify model's momentum according to learning rate, which could make the model converge in a faster way.
@@ -198,7 +198,7 @@ custom_hooks = [
 
 ### Modify default runtime hooks
 
-There are some common hooks that are not registerd through `custom_hooks`, they are
+There are some common hooks that are not registered through `custom_hooks`, they are
 
 - log_config
 - checkpoint_config
