+++ tools/model_converters/mit2mmseg.py
@@ -14,14 +14,14 @@ def convert_mit(ckpt):
     for k, v in ckpt.items():
         if k.startswith('head'):
             continue
-        # patch embedding convertion
+        # patch embedding conversion
         elif k.startswith('patch_embed'):
             stage_i = int(k.split('.')[0].replace('patch_embed', ''))
             new_k = k.replace(f'patch_embed{stage_i}', f'layers.{stage_i-1}.0')
             new_v = v
             if 'proj.' in new_k:
                 new_k = new_k.replace('proj.', 'projection.')
-        # transformer encoder layer convertion
+        # transformer encoder layer conversion
         elif k.startswith('block'):
             stage_i = int(k.split('.')[0].replace('block', ''))
             new_k = k.replace(f'block{stage_i}', f'layers.{stage_i-1}.1')
@@ -45,7 +45,7 @@ def convert_mit(ckpt):
                 new_k = new_k.replace('dwconv.dwconv.', '1.')
                 new_k = new_k.replace('fc2.', '4.')
                 string += f'{new_k} {v.shape}-{new_v.shape}'
-        # norm layer convertion
+        # norm layer conversion
         elif k.startswith('norm'):
             stage_i = int(k.split('.')[0].replace('norm', ''))
             new_k = k.replace(f'norm{stage_i}', f'layers.{stage_i-1}.2')
